com.speruri.java.cryptography/src/main/java/com/speruri/java/cryptography/asymmetric/AsymmetricEncryptionUtils.java
package com.speruri.java.cryptography.asymmetric;

import java.security.KeyPair;
import java.security.KeyPairGenerator;
import java.security.PrivateKey;
import java.security.PublicKey;
import java.security.SecureRandom;

import javax.crypto.Cipher;

public class AsymmetricEncryptionUtils {

	private static String RSA = "RSA";

	public static KeyPair generateRSAKeypair() throws Exception {
		SecureRandom secureRandom = new SecureRandom();
		KeyPairGenerator keyPairGenerator = KeyPairGenerator.getInstance(RSA);
		keyPairGenerator.initialize(4096, secureRandom);
		KeyPair keyPair = keyPairGenerator.generateKeyPair();
		return keyPair;
	}

	public static byte[] performRSAEncryption(String plainText, PrivateKey privateKey) throws Exception {
		Cipher cipher = Cipher.getInstance(RSA);
		cipher.init(Cipher.ENCRYPT_MODE, privateKey);
		return cipher.doFinal(plainText.getBytes());
	}

	public static String performRSADecryption(byte[] encryptedBytes, PublicKey publicKey) throws Exception {
		Cipher cipher = Cipher.getInstance(RSA);
		cipher.init(Cipher.DECRYPT_MODE, publicKey);
		System.out.println(cipher.getProvider().getName());
		byte[] plainBytes = cipher.doFinal(encryptedBytes);
		return new String(plainBytes);
	}

}
